requirements.txt
absl-py==2.3.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
immutabledict==4.2.2
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
ortools==9.14.6206
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.31.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
wheel==0.45.1
streamlit>=1.30
filelock>=3.12
